fix(tanstack-vue): move unit tests
- Introduced comprehensive unit tests for the `asTanStackEndpoint`, `enhanceEndpointWithQuery`, and `enhanceEndpointWithMutation` functions.
- Verified the correct structure and functionality of query and mutation options.
- Ensured proper handling of URL parameters and input in query and mutation functions.
- Added tests for error handling in both query and mutation scenarios.
- Created a mock API client to facilitate testing of endpoint enhancements.
